Add Rust language support to skiboot
This is the result of a small experiment to add Rust code to skiboot
for fun and profit, it may or may not be useful for anything at
all. The Makefiles have been setup to build rust code and a small
number of standard library functions have been ported (see librust/).

Building requires a Rust compiler in your path along with the compiler
source which is assumed to be in the "rust" directory one level up from
the skiboot source. If not it may be specified as a path relative to
the skiboot source with:

RUSTSRC=$(realpath --relative-to=$PWD <rust src dir>) make

This will build the base Rust libraries along with skiboot which may
then be run on your favourite hardware or system emulator. It has been
tested with rustc 1.8.0-nightly (094c5b0d6 2016-01-31).

Signed-off-by: Alistair Popple <alistair@popple.id.au>