PCI: Move PHB lock to generic layer
All kinds of PHBs are maintaining a spinlock. At meanwhile, the
spinlock is acquired or released by backends for phb_ops->lock()
or phb_ops->unlock(). There're no difference of the logic on all
kinds of PHBs. So it's reasonable to maintain the lock in the
generic layer (struct phb).

This moves lock from specific PHB to generic one. Also, two inline
functions (phb_{lock, unlock}()) are introduced to acquire/release
it.

Signed-off-by: Gavin Shan <gwshan@linux.vnet.ibm.com>