PCI: Init phb->devices in backends
Currently, the list head (phb->devices) is initialized when the
PHB is registered. Other fields in @phb is initialized in the
backends when it's allocated.

This moves the initialization (phb->device) from pci_register_phb()
to backends. At the same time, the list head for PCI virtual devices
is initialized.

Signed-off-by: Gavin Shan <gwshan@linux.vnet.ibm.com>