npu2: Fix BAR mapping for multiple chips
NPU2 BARs were being assigned and tracked with a global static
array. This worked fine when there was only a single chip/NPU2 in the
system however multiple chips results in the a shared data structure
for BAR management which results in multiple chips getting assigned
the same BAR addresses and other incorrect sharing of BAR properties.

This patch splits the static and dynamic BAR configuration and stores
the dynamic configuration in the per-NPU2 data structure.

Signed-off-by: Alistair Popple <alistair@popple.id.au>